Finished roles tests

+++ elements/role/role.model.ts
@@ -79,6 +79,8 @@ const RoleSchema = new mongoose.Schema<IRole>(
 RoleSchema.pre("deleteOne", async function (next) {
   const role: IRole = (await this.model.findOne(this.getQuery())) as IRole;
 
+  if (!role) return;
+
   await entityPermissionRepository.deleteByIds(
     role.entityPermissions.map((p) => p._id.toString())
   );
